+++ packages/e2e-tests/plugins/interactive-blocks/directive-on-document/render.php
@@ -8,10 +8,21 @@
 gutenberg_enqueue_module( 'directive-on-document-view' );
 ?>
 
-<div data-wp-interactive='{ "namespace": "directive-on-document" }' data-wp-context='{"isVisible":true}'>
-	<button data-wp-on--click="actions.visibilityHandler" data-testid="visibility">Switch visibility</button>
-	<div data-wp-show-mock="context.isVisible">
-		<div data-wp-on-document--keydown="callbacks.keydownHandler">
+<div data-wp-interactive='{ "namespace": "directive-on-document" }'>
+	<button 
+		data-testid="visibility"
+		data-wp-on--click="actions.visibilityHandler"
+	>
+		Switch visibility
+	</button>
+
+	<div data-wp-text="state.isEventAttached" data-testid="isEventAttached">no</div>
+
+	<div data-wp-show-mock="state.isVisible">
+		<div
+			data-wp-on-document--keydown="callbacks.keydownHandler"
+			data-wp-init="callbacks.init"
+		>
 			<p data-wp-text="state.counter" data-testid="counter">0</p>
 		</div>
 	</div>